feat: handled navigation from movieDetail
